Set security timed cb xevent queue in constructor
That makes for a more intuitive interface.  It requires updating all the
call-sites, which in particular affects the loading of security plugins
in tests.

The loader function used by the tests now always passes a "gv" pointer
with an event queue in it to the plugins.  The authentication plug-in
did some test-specific things when no "gv" was pssed in, that has is now
done by explicitly setting the relevant flag in "gv".

Signed-off-by: Erik Boasson <eb@ilities.com>